add a reset menu item to reset the initial state of the program
We add a menu item "Reset" by clicking into the menu bar in the
designer. In the action editor window we right click on the new action
"Reset" and then "Go to slot..." which opens up a window with all
possible actions. We choose "triggered()" and then fill the code into
the newly created action stub.