add an icon for reset to the toolbar
This way the reset menu item is easier to access.

We first drew an svg image using inkscape. Then we right clicked on the
project and choose "Add new...". In the new window we choose "Qt
resource file" and named it "icons.qrc". We then added the already
existing icon file with "Add existing files...". Now we doubled clicked
the new resource file, added a new prefix and then added the icon.

To add the action to the toolbar we dragged it in the designer from the
action editor to the empty toolbar on top of the window. We then edited
the action in the action editor and selected the icon from the resource
file.